tests/wpt/tests/css/css-scroll-anchoring/descend-into-container-with-overflow.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>

body { height: 4000px; }
#outer { width: 300px; }
#zeroheight { height: 0px; }
#changer { height: 100px; background-color: red; }
#bottom { margin-top: 600px; }

</style>
<div id="outer">
  <div id="zeroheight">
    <div id="changer"></div>
    <div id="bottom">bottom</div>
  </div>
</div>
<script>

// Tests that the anchor selection algorithm descends into zero-height
// containers that have overflowing content.

test(() => {
  document.scrollingElement.scrollTop = 200;
  document.querySelector("#changer").style.height = "200px";
  assert_equals(document.scrollingElement.scrollTop, 300);
}, "Zero-height container with visible overflow.");

</script>
